remove old loaders
