+++ src/levanter/data/__init__.py
@@ -1,6 +1,5 @@
 from ._preprocessor import BatchProcessor
 from .dataset import Dataset, ShardableDataset, ShuffleDataset
-from .loader import BatchLoader, ReplicatedBatchLoader, ShardedBatchLoader
 from .sharded_dataset import ShardedDataSource, datasource_from_hf, datasource_from_jsonl
 from .utils import batched
 
@@ -10,9 +9,6 @@ __all__ = [
     "Dataset",
     "ShardableDataset",
     "ShuffleDataset",
-    "BatchLoader",
-    "ReplicatedBatchLoader",
-    "ShardedBatchLoader",
     "ShardedDataSource",
     "datasource_from_hf",
     "datasource_from_jsonl",
